+++ CHANGELOG.md
@@ -54,6 +54,7 @@
 - HTML: minor improvements to Album Art help and updated Biography default settings (to match latest versions).
 - HTML: minor improvements and cleanup at HTML options panel.
 - Syntax: updated syntax docs with relevant changes and examples.
+- Debug: improvements to profiling logging for library loading.
 - Code cleanup.
 ### Removed
 ### Fixed